Move device_registry cleaning to __init__.py
